Dispatch WS commands as bounded tracked tasks, cancel on disconnect
