+++ docs/ARCHITECTURE.md
@@ -97,7 +97,9 @@ esphome_device_builder/
 
 ### Command dispatch concurrency
 
-The `/ws` message loop **awaits ordinary commands inline**, so commands on a single connection execute in submission order — two dependent mutations (e.g. back-to-back `set_preferences`, or `auth` before a privileged command) can't interleave, and ordinary commands spawn no background task. Only **streaming** handlers — the long-lived subscriptions that park for the connection lifetime (`subscribe_events`, `firmware/follow_job(s)`, `devices/logs`, `devices/subscribe_reachability`) — are run as spawned tasks, marked with `@api_command(..., streaming=True)` and collected into `DeviceBuilder.streaming_commands`. The per-connection streaming task set is bounded by `_MAX_CONCURRENT_STREAMS` (32); a stream-open frame past the cap is refused with `RATE_LIMITED` rather than fanning out unbounded server-side tasks. This mirrors the bounded outbound side (the 4000-slot `stream_events` queue) on the inbound command path (advisory GHSA-mg7m-j658-c6r9).
+The `/ws` message loop dispatches each inbound command as a **tracked, per-connection task** rather than awaiting it inline, so a slow handler (e.g. an executor-bound `validate_config`) doesn't head-of-line-block faster commands like `ping` / `devices/list` on the same connection. Every task is cancelled on disconnect (`WebSocketClient.cleanup`) — the Home Assistant `websocket_api` pattern. Fan-out is bounded per connection so a client spamming frames can't grow unbounded tasks/memory (advisory GHSA-mg7m-j658-c6r9): ordinary commands are capped at `_MAX_CONCURRENT_COMMANDS` (64), and **streaming** handlers — the long-lived subscriptions that park for the connection lifetime (`subscribe_events`, `firmware/follow_job(s)`, `devices/logs`, `devices/subscribe_reachability`), marked `@api_command(..., streaming=True)` and collected into `DeviceBuilder.streaming_commands` — get a tighter cap of `_MAX_CONCURRENT_STREAMS` (32). A frame past either cap is refused with `RATE_LIMITED`. This mirrors the bounded outbound side (the 4000-slot `stream_events` queue) on the inbound command path.
+
+The per-frame parsing, auth-flag check, and streaming-vs-ordinary decision still run inline in receipt order; only the handler *body* runs concurrently. Same-connection commands are therefore no longer forced into strict submission order — the guarantees that previously leaned on it are enforced elsewhere: pre-auth gating by the `_PRE_AUTH_COMMANDS` check (a privileged frame before `auth` is rejected on the authenticated flag regardless of dispatch order), and lost-update mutation races by the `metadata_transaction` helpers (`preferences_transaction` / `labels_transaction`).
 
 ## Event bus
